gpu: ion: Fix incorrect argument to map call
An incorrect argument is passed to a local ion heap
function call. This is causing incorrect flags to be
passed to iommu map call. The side effect is that the
flags passed to the iommu map call is defaulted to
read-write and uncached.

Change-Id: Ia2a7e64818dc8f251dadfb545f9b218a5e3cfc42
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
(cherry picked from commit 3d79eec14b8fdb99832a907a0ae29042eaba7f07)